send kafka message after update request

+++ src/main/java/com/sd/devicemanagement/dto/DeviceUpdateDTO.java
@@ -3,6 +3,6 @@ package com.sd.devicemanagement.dto;
 public record DeviceUpdateDTO(
         String username,
         String deviceName,
-        float maxConsumption
+        double maxConsumption
 ) {
 }
